Towards a real spacecraft in gaia sky

+++ core/src/gaia/cu9/ari/gaiaorbit/interfce/AbstractGui.java
@@ -49,8 +49,8 @@ public abstract class AbstractGui implements IObserver, IGui {
     }
 
     @Override
-    public void update(float dt) {
-        ui.act(dt);
+    public void update(double dt) {
+        ui.act((float) dt);
     }
 
     @Override
