+++ core/src/gaia/cu9/ari/gaiaorbit/interfce/FullGui.java
@@ -261,8 +261,8 @@ public class FullGui extends AbstractGui {
     }
 
     @Override
-    public void update(float dt) {
-        ui.act(dt);
+    public void update(double dt) {
+        ui.act((float) dt);
         notificationsInterface.update();
     }
