+++ core/src/gaia/cu9/ari/gaiaorbit/render/IAtmosphereRenderable.java
@@ -4,6 +4,7 @@ import com.badlogic.gdx.graphics.g3d.ModelBatch;
 
 /**
  * To be implemented by all entities wanting to render an atmosphere.
+ * 
  * @author Toni Sagrista
  *
  */
@@ -11,10 +12,15 @@ public interface IAtmosphereRenderable extends IRenderable {
 
     /**
      * Renders the atmosphere.
-     * @param modelBatch The model batch to use.
-     * @param alpha The opacity.
-     * @param t The time in seconds since the start.
-     * @param atm Render atmosphere.
+     * 
+     * @param modelBatch
+     *            The model batch to use.
+     * @param alpha
+     *            The opacity.
+     * @param t
+     *            The time in seconds since the start.
+     * @param atm
+     *            Render atmosphere.
      */
-    public void render(ModelBatch modelBatch, float alpha, float t, boolean atm);
+    public void render(ModelBatch modelBatch, float alpha, double t, boolean atm);
 }
