+++ core/src/gaia/cu9/ari/gaiaorbit/render/SGR.java
@@ -19,7 +19,7 @@ public class SGR extends SGRAbstract implements ISGR {
     }
 
     @Override
-    public void render(SceneGraphRenderer sgr, ICamera camera, float t, int rw, int rh, FrameBuffer fb, PostProcessBean ppb) {
+    public void render(SceneGraphRenderer sgr, ICamera camera, double t, int rw, int rh, FrameBuffer fb, PostProcessBean ppb) {
         boolean postproc = postprocessCapture(ppb, fb, rw, rh);
 
         extendViewport.setCamera(camera.getCamera());
