+++ core/src/gaia/cu9/ari/gaiaorbit/render/system/FontRenderSystem.java
@@ -28,66 +28,64 @@ public class FontRenderSystem extends AbstractRenderSystem {
     private Comparator<IRenderable> comp;
 
     public FontRenderSystem(RenderGroup rg, int priority, float[] alphas, SpriteBatch batch) {
-	super(rg, priority, alphas);
-	this.batch = batch;
+        super(rg, priority, alphas);
+        this.batch = batch;
 
-	// Init comparator
-	comp = new DistToCameraComparator<IRenderable>();
+        // Init comparator
+        comp = new DistToCameraComparator<IRenderable>();
     }
 
-    public FontRenderSystem(RenderGroup rg, int priority, float[] alphas, SpriteBatch batch,
-	    ShaderProgram shaderProgram) {
-	this(rg, priority, alphas, batch);
-	this.shaderProgram = shaderProgram;
+    public FontRenderSystem(RenderGroup rg, int priority, float[] alphas, SpriteBatch batch, ShaderProgram shaderProgram) {
+        this(rg, priority, alphas, batch);
+        this.shaderProgram = shaderProgram;
 
-	// 3D font
-	Texture texture3d = new Texture(Gdx.files.internal("font/main-font.png"), true);
-	texture3d.setFilter(TextureFilter.Linear, TextureFilter.Linear);
-	font3d = new BitmapFont(Gdx.files.internal("font/main-font.fnt"), new TextureRegion(texture3d), false);
+        // 3D font
+        Texture texture3d = new Texture(Gdx.files.internal("font/main-font.png"), true);
+        texture3d.setFilter(TextureFilter.Linear, TextureFilter.Linear);
+        font3d = new BitmapFont(Gdx.files.internal("font/main-font.fnt"), new TextureRegion(texture3d), false);
 
-	// 2D font
-	Texture texture2d = new Texture(Gdx.files.internal("font/font2d.png"), true);
-	texture2d.setFilter(TextureFilter.Linear, TextureFilter.Linear);
-	font2d = new BitmapFont(Gdx.files.internal("font/font2d.fnt"), new TextureRegion(texture2d), false);
+        // 2D font
+        Texture texture2d = new Texture(Gdx.files.internal("font/font2d.png"), true);
+        texture2d.setFilter(TextureFilter.Linear, TextureFilter.Linear);
+        font2d = new BitmapFont(Gdx.files.internal("font/font2d.fnt"), new TextureRegion(texture2d), false);
 
     }
 
     @Override
-    public void renderStud(Array<IRenderable> renderables, ICamera camera, float t) {
-	renderables.sort(comp);
-	batch.begin();
-	int size = renderables.size;
-	if (shaderProgram == null) {
-	    for (int i = 0; i < size; i++) {
-		IAnnotationsRenderable s = (IAnnotationsRenderable) renderables.get(i);
-		// Render sprite
-		s.render(batch, camera, getAlpha(s));
-	    }
-	} else {
-	    float lalpha = alphas[ComponentType.Labels.ordinal()];
-	    font3d.getData().setScale(0.6f);
-	    for (int i = 0; i < size; i++) {
-		I3DTextRenderable s = (I3DTextRenderable) renderables.get(i);
+    public void renderStud(Array<IRenderable> renderables, ICamera camera, double t) {
+        renderables.sort(comp);
+        batch.begin();
+        int size = renderables.size;
+        if (shaderProgram == null) {
+            for (int i = 0; i < size; i++) {
+                IAnnotationsRenderable s = (IAnnotationsRenderable) renderables.get(i);
+                // Render sprite
+                s.render(batch, camera, getAlpha(s));
+            }
+        } else {
+            float lalpha = alphas[ComponentType.Labels.ordinal()];
+            font3d.getData().setScale(0.6f);
+            for (int i = 0; i < size; i++) {
+                I3DTextRenderable s = (I3DTextRenderable) renderables.get(i);
 
-		// Regular mode, we use 3D distance field font
-		I3DTextRenderable lr = (I3DTextRenderable) s;
-		shaderProgram.setUniformf("a_labelAlpha", (lr.isLabel() || camera.getCurrent() instanceof FovCamera
-			? lalpha * lr.textColour()[3] : lr.textColour()[3]));
-		shaderProgram.setUniformf("a_componentAlpha", getAlpha(s));
-		// Font opacity multiplier
-		shaderProgram.setUniformf("u_opacity", 0.7f);
+                // Regular mode, we use 3D distance field font
+                I3DTextRenderable lr = (I3DTextRenderable) s;
+                shaderProgram.setUniformf("a_labelAlpha", (lr.isLabel() || camera.getCurrent() instanceof FovCamera ? lalpha * lr.textColour()[3] : lr.textColour()[3]));
+                shaderProgram.setUniformf("a_componentAlpha", getAlpha(s));
+                // Font opacity multiplier
+                shaderProgram.setUniformf("u_opacity", 0.7f);
 
-		s.render(batch, shaderProgram, font3d, font2d, camera);
-	    }
-	}
-	batch.end();
+                s.render(batch, shaderProgram, font3d, font2d, camera);
+            }
+        }
+        batch.end();
 
     }
 
     @Override
     public void resize(int w, int h) {
-	super.resize(w, h);
-	batch.getProjectionMatrix().setToOrtho2D(0, 0, w, h);
+        super.resize(w, h);
+        batch.getProjectionMatrix().setToOrtho2D(0, 0, w, h);
     }
 
 }
