+++ core/src/gaia/cu9/ari/gaiaorbit/render/system/ParticleGroupRenderSystem.java
@@ -68,7 +68,7 @@ public class ParticleGroupRenderSystem extends ImmediateRenderSystem implements 
     }
 
     @Override
-    public void renderStud(Array<IRenderable> renderables, ICamera camera, float t) {
+    public void renderStud(Array<IRenderable> renderables, ICamera camera, double t) {
         renderables.sort(comp);
         if (renderables.size > 0) {
             for (IRenderable renderable : renderables) {
