+++ core/src/gaia/cu9/ari/gaiaorbit/scenegraph/LightBeam.java
@@ -73,7 +73,7 @@ public class LightBeam extends ModelBody {
     }
 
     @Override
-    public void render(ModelBatch modelBatch, float alpha, float t) {
+    public void render(ModelBatch modelBatch, float alpha, double t) {
         mc.touch();
         mc.setTransparency(alpha * opacity);
         modelBatch.render(mc.instance, mc.env);
